+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Add a performance timer for monitoring
 * Implement FocusIn/FocusOut inputs (#367)
 * Add lua window to manage scripts (#519)
+* Update the seek frame during rewind (#530)
 
 ### Changed